improve logging (#120)
