+++ deeppavlov_agent/core/__init__.py
@@ -1,5 +1,1 @@
-from .log import init_logger
-
 STATE_API_VERSION = "0.13.0"
-
-init_logger()
