--- deeppavlov_agent/log_config.yml
@@ -1,27 +1,0 @@
-version: 1
-disable_existing_loggers: false
-loggers:
-  core:
-    level: 'INFO'
-    propagate: true
-    handlers: ['stderr']
-formatters:
-  default:
-    format: "%(asctime)s.%(msecs)d %(levelname)s in '%(name)s'['%(module)s'] at line %(lineno)d: %(message)s"
-    datefmt: "%Y-%m-%d %H:%M:%S"
-handlers:
-  file:
-    class: 'logging.FileHandler'
-    level: 'DEBUG'
-    formatter: 'default'
-    filename: '~/dp_agent.log'
-  stdout:
-    class: 'logging.StreamHandler'
-    level: 'DEBUG'
-    formatter: 'default'
-    stream: 'ext://sys.stdout'
-  stderr:
-    class: 'logging.StreamHandler'
-    level: 'DEBUG'
-    formatter: 'default'
-    stream: 'ext://sys.stderr'
