+++ deeppavlov_agent/run.py
@@ -1,14 +1,15 @@
+import sys
 import logging
 
 import hydra
 from omegaconf import DictConfig
 
+import deeppavlov_agent.log as log
 from .run_cmd import run_cmd
 from .run_http import run_http
 from .run_tg import run_telegram
 
-
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("run")
 
 CHANNELS = {
     "cmd": run_cmd,
@@ -19,13 +20,13 @@ CHANNELS = {
 
 @hydra.main(config_path=".", config_name="settings")
 def main(cfg: DictConfig):
-    try:
-        run_channel = CHANNELS[cfg.agent.channel]
-        run_channel(cfg)
-    except KeyError:
-        logger.error(
-            f"agent.channel value must be one of: {', '.join(CHANNELS.keys())} (not {cfg.agent.channel})"
-        )
+    with log.setup():
+        try:
+            run_channel = CHANNELS[cfg.agent.channel]
+            run_channel(cfg)
+        except Exception as e:
+            logger.exception(e)
+            sys.exit(-1)
 
 
 if __name__ == "__main__":
